requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# LLM and AI
openai==1.3.8
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2

# Vector Databases
pinecone-client==2.2.4
weaviate-client==3.25.3
chromadb==0.4.18

# Data Processing
pandas==2.1.4
numpy==1.25.2
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.0
openpyxl==3.1.2
beautifulsoup4==4.12.2
requests==2.31.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
redis==5.0.1

# MLOps
mlflow==2.8.1
wandb==0.16.1
dvc==3.37.0
evidently==0.4.7

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# API Documentation
python-dotenv==1.0.0
jinja2==3.1.2
